Rename stages/ -> steps/ to match ML pipeline vocabulary
"Steps" is the standard term across the ML ecosystem for the units a pipeline
runs in sequence (scikit-learn Pipeline(steps=[...]), ZenML @step, SageMaker
Pipelines, MLflow). It pairs cleanly with indexing.py as the pipeline that runs
parsing -> chunking -> summarize in order.

- git mv rag/stages -> rag/steps
- update imports in indexing.py and __init__.py (aliases *_stage -> *_step)
- update docstrings (stage -> step)
